reorg functions directory (#6393)
