+++ langchain/chains/__init__.py
@@ -24,6 +24,7 @@ from langchain.chains.mapreduce import MapReduceChain
 from langchain.chains.moderation import OpenAIModerationChain
 from langchain.chains.natbot.base import NatBotChain
 from langchain.chains.openai_functions import (
+    create_citation_fuzzy_match_chain,
     create_extraction_chain,
     create_extraction_chain_pydantic,
     create_tagging_chain,
@@ -93,4 +94,5 @@ __all__ = [
     "create_tagging_chain",
     "create_tagging_chain_pydantic",
     "load_chain",
+    "create_citation_fuzzy_match_chain",
 ]
