+++ langchain/chains/openai_functions/__init__.py
@@ -1,0 +1,19 @@
+from langchain.chains.openai_functions.citation_fuzzy_match import (
+    create_citation_fuzzy_match_chain,
+)
+from langchain.chains.openai_functions.extraction import (
+    create_extraction_chain,
+    create_extraction_chain_pydantic,
+)
+from langchain.chains.openai_functions.tagging import (
+    create_tagging_chain,
+    create_tagging_chain_pydantic,
+)
+
+__all__ = [
+    "create_tagging_chain",
+    "create_tagging_chain_pydantic",
+    "create_extraction_chain_pydantic",
+    "create_extraction_chain",
+    "create_citation_fuzzy_match_chain",
+]
